requirements.txt
python-dotenv==0.15.0
python-telegram-bot==11.1.0
redis==3.2.1
requests==2.25.1